Use size_t.

+++ teasort.cc
@@ -12,7 +12,7 @@ template <typename T>
 struct Vertex
 {
   typedef std::vector<Vertex*> Neighbours;
-  unsigned int id;
+  size_t id;
   T value;
   bool visited;
   Neighbours neighbours;
@@ -26,17 +26,17 @@ struct Graph
   typedef typename std::vector<Vertex<T>>::iterator iterator;
   size_t size;
   std::vector<Vertex<T>> vertices;
-  Graph(unsigned int nsize)
+  Graph(size_t nsize)
   : vertices(nsize)
   {
     size = nsize;
-    for (int i = 0; i < size; i++)
+    for (size_t i = 0; i < size; i++)
       vertices[i].id = i;
   }
   inline iterator begin(void) { return vertices.begin(); }
   inline iterator end(void) { return vertices.end(); }
-  inline Vertex<T> &operator[] (unsigned int n) { return vertices[n]; }
-  inline void add_edge(unsigned int x, unsigned int y)
+  inline Vertex<T> &operator[] (size_t n) { return vertices[n]; }
+  inline void add_edge(size_t x, size_t y)
   {
     assert(x < size && y < size);
     vertices[x].neighbours.push_back(&vertices[y]);
@@ -71,7 +71,7 @@ std::vector<Vertex<T>*> dfs(Graph<T> &graph)
 
 uintmax_t teasort(int *p, int *q)
 {
-  int n = q - p;
+  size_t n = q - p;
   uintmax_t m = static_cast<uintmax_t>(2) * n * std::ilogb((double)n);
 
   static std::random_device rdev;
@@ -79,13 +79,13 @@ uintmax_t teasort(int *p, int *q)
   std::uniform_int_distribution<> rand(0, n - 1);
 
   Graph<int> graph(n);
-  for (int i = 0; i < n; i++)
+  for (size_t i = 0; i < n; i++)
     graph[i].value = p[i];
 
-  for (int i = 0; i < m; i++)
+  for (size_t i = 0; i < m; i++)
   {
-    int x = rand(rng);
-    int y = rand(rng);
+    size_t x = rand(rng);
+    size_t y = rand(rng);
     if (p[y] > p[x])
       std::swap(x, y);
     graph.add_edge(x, y);
@@ -97,11 +97,11 @@ uintmax_t teasort(int *p, int *q)
 
   std::vector<int> t(n + 1);
   t[0] = std::numeric_limits<int>::min();
-  for (int i = 0; i < n; i++)
+  for (size_t i = 0; i < n; i++)
     t[i + 1] = result[i]->value;
-  for (int i = 2; i <= n; i++)
+  for (size_t i = 2; i <= n; i++)
   {
-    int j = i;
+    size_t j = i;
     int v = t[i];
     while (m++, t[j - 1] > v) {
       t[j] = t[j - 1];
@@ -119,7 +119,7 @@ uintmax_t teasort(int *p, int *q)
 
 int main(void)
 {
-  for (int n = TEST_MIN; ; n *= 2)
+  for (size_t n = TEST_MIN; ; n *= 2)
   {
     std::vector<int> p(n);
     std::iota(p.begin(), p.end(), 1);
@@ -127,7 +127,7 @@ int main(void)
     for (int i = 0; i < TEST_ITER; i++) {
       std::random_shuffle(p.begin(), p.end());
       cost += teasort(&p[0], &p[0] + n);
-      for (int k = 0; k < n; k++)
+      for (size_t k = 0; k < n; k++)
         assert(p[k] == k + 1);
     }
     double fcost = (double)cost / (double)TEST_ITER;
